+++ perf/__init__.py
@@ -1,6 +1,6 @@
 from __future__ import division, print_function, absolute_import
 
-__version__ = '0.7.13'
+__version__ = '0.8.0'
 
 # Clocks
 try: